third_party/WebKit/LayoutTests/intersection-observer/iframe-no-root.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
pre, #log {
  position: absolute;
  top: 0;
  left: 200px;
}
.spacer {
  height: 2000px;
}
iframe {
  height: 100px;
  width: 150px;
}
</style>

<div class="spacer"></div>
<iframe id="target-iframe" src="resources/iframe-no-root-subframe.html"></iframe>
<div class="spacer"></div>

<script>
var iframe = document.getElementById("target-iframe");
var target;
var entries = [];

iframe.onload = function() {
  runTestCycle(function() {
    assert_true(!!iframe, "iframe exists");
    assert_greater_than(window.innerWidth, 200, "Window must be at least 200 pixels wide.");
    assert_greater_than(window.innerHeight, 200, "Window must be at least 200 pixels high.");
    assert_less_than(window.innerHeight, 2000, "Window must be no more than 2000 pixels high.");

    target = iframe.contentDocument.getElementById("target");
    assert_true(!!target, "Target element exists.");
    var observer = new IntersectionObserver(function(changes) {
      entries = entries.concat(changes);
      console.log(changes[0]);
    });
    observer.observe(target);
    entries = entries.concat(observer.takeRecords());
    assert_equals(entries.length, 0, "No initial notifications.");
    runTestCycle(step0, "First rAF.");
  }, "Observer with the implicit root; target in a same-origin iframe.");
};

function step0() {
  var scrollTop = 2100 - document.documentElement.clientHeight;
  document.scrollingElement.scrollTop = scrollTop;
  runTestCycle(step1, "document.scrollingElement.scrollTop = " + scrollTop);
  checkLastEntry(entries, 0, [8, 108, 208, 308, 0, 0, 0, 0].concat(topBounds()).concat(false));
}

function step1() {
  iframe.contentDocument.scrollingElement.scrollTop = 250;
  runTestCycle(step2, "iframe.contentDocument.scrollingElement.scrollTop = 250");
  assert_equals(entries.length, 1, "entries.length == 1");
}

function step2() {
  document.scrollingElement.scrollTop = 0;
  runTestCycle(step3, "document.scrollingElement.scrollTop = 0");
  checkLastEntry(entries, 1, [8, 108, -42, 58, 8, 108, 0, 58].concat(topBounds()).concat(true));
}

function step3() {
  checkLastEntry(entries, 2, [8, 108, -42, 58, 0, 0, 0, 0].concat(topBounds()).concat(false));
}
</script>
